requirements.txt
mkdocs>=1.6.0
mkdocs-material>=9.5.0
mkdocs-macros-plugin>=1.0.5
mkdocs-gen-files>=0.5.0
PyYAML>=6.0.1
beautifulsoup4>=4.12.3
lxml>=5.2.2
